Update Visual scripts
Tried to fix the problem mentioned in class where the enemies don't destroy if you add the prefab to a tile in the maze. It didn't work. some kind of error with the "get position" script and nav mesh idk. also still have no clue why it's so pixelated, maybe my laptop's gpu isn't strong enough